Add Support for Querying Grid's Input Class
Specifically, we add support for querying an EclipseGrid instance as
to whether or not the input grid is radial (EclipseGrid::isRadial())
in which case clients should request radial permeability tensor
keywords or if it is Cartesian (roughly, whether or not it uses
corner-point specification).

We furthermore fold the "is full circle" predicate into a bitmask
that answers the above questions.