SummaryConfig: Recognise Subset of Segment Vectors
This commit extends the SummaryConfig class to recognise a small
subset of segment-related summary vectors.  In particular, this
brings support for 'SGFR', 'SOFR', 'SPR', and 'SWFR'--at least in a
restricted sense.  We do not yet support cases like

    SOFR
    /

which designates all segments in all wells at all times.

The unit testing is presently minimal and must be expanded before
this is ready for inclusion into master.  In particular, we only
check that cases like

    SOFR
      'PROD01' 1 /
    /

generate the expected summary vector nodes (class SummaryNode).