Add Conditional Low-Pressure Padding of Gas Property Tables
This commit adds logic to extend gas property tables entered using
the PVDG or PVTG to low pressure values if needed.  When we compute
well production potentials (e.g., summary vectors WxPP) and the
well's bottom-hole pressure limit is defaulted we end up using
pressures in the order of one atmosphere (1 atm) and this, in turn,
may generate negative formation volume factors (mass densities)
and/or phase viscosities.  The problem is most pronounced for
producing wells in history matching mode (keyword WCONHIST), but we
should have guards in place to avoid negative densities regardless.

If needed we extend the tables down to 1 barsa and also insert a
"limiting" pressure value PLim corresponding to a maximum FVF value
in the range [1, 2].  In this case we create new input tables and
redo the property table initialisation with this extended input table.