Lay Foundation for More Strict Array Operations
This commit makes the backend implementations of ADD, MULTIPLY, &c
capable of halting processing if we encounter any undefined source
elements.  We nevertheless maintain the definition of "defined" as

    value::has_value()

(from the value_status.hpp header), at least for the time being, as
changing the definition requires making more parts of the array
processing aware of the currently active section.

The new helper function reject_undefined_operation() will throw an
OpmInputError exception with (hopefully) enough context that fixing
the input will be easier.

While here, make the apply() function private to the FieldProps.cpp
implementation file.  It was already marked 'static' and does not
need access to any of the FieldProps data members.